+++ CHANGELOG.md
@@ -27,7 +27,7 @@ For more information on SemVer, please visit http://semver.org/.
 
 * Added `_WPDKPost::metaValue()` and `_WPDKPost::metaValues()` utility methods
 * Improved wpdk_is_user_logged_in shortcode with roles, caps, emails and ids attributes
-
+* Minor stable enhancements
 
 
 ## Version 1.3.0