+++ CHANGELOG.md
@@ -25,6 +25,7 @@ For more information on SemVer, please visit http://semver.org/.
 
 #### Enhancements
 
+* Added `WPDKMenu::addSubMenusAt()` utility method
 * Added `WPDKHTMLTagImg` class
 * Added `_WPDKPost::metaValue()` and `_WPDKPost::metaValues()` utility methods
 * Improved wpdk_is_user_logged_in shortcode with roles, caps, emails and ids attributes