+++ CHANGELOG.md
@@ -27,12 +27,11 @@ For more information on SemVer, please visit http://semver.org/.
 
 * Several improvements (no conflict) to align to Bootstrap v3.0.0
 * Added several utility static methods in `WPDKHTMLTag`
-* Added `WPDKGlyphIcons`
-* Introducing WPDK fonts ( see assets/fonts/ folder for detail )
+* Introducing `WPDKGlyphIcons` ( see assets/fonts/ folder for detail )
 * Added `WPDKMenu::addSubMenusAt()` utility method
 * Added `WPDKHTMLTagImg` class
 * Added `_WPDKPost::metaValue()` and `_WPDKPost::metaValues()` utility methods
-* Improved wpdk_is_user_logged_in shortcode with roles, caps, emails and ids attributes
+* Improved `wpdk_is_user_logged_in` shortcode with roles, caps, emails and ids attributes
 * Minor stable enhancements